backend/requirements.txt
# Core
fastapi==0.109.0
uvicorn[standard]==0.27.0
gunicorn==21.2.0
python-multipart==0.0.6
python-dotenv==1.0.0

# Grok & AI
httpx==0.26.0
openai==1.9.0
langchain==0.1.3
langchain-community==0.0.38
langchain-openai==0.0.3
faiss-cpu==1.7.4
tiktoken==0.5.2

# Document Processing
pypdf==3.17.4
unstructured[local-inference]==0.11.6
python-magic-bin==0.4.14  # For Windows
pytesseract==0.3.10

# Security
cryptography==41.0.7
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pydantic[email]==2.5.3

# Database & Caching
sqlalchemy==2.0.25
redis==5.0.1
alembic==1.13.1
psycopg2-binary==2.9.9

# Monitoring & Analytics
prometheus-client==0.19.0
sentry-sdk==1.39.2

# Utilities
tenacity==8.2.3
python-dateutil==2.8.2
pytz==2023.3
validators==0.22.0

# WebSocket
websockets==12.0

# Testing
pytest==7.4.4
pytest-asyncio==0.23.3